+++ framework/common/time_util.h
@@ -24,22 +24,20 @@ void time_util_init(void);
 
 /**
  * @brief  获取自系统启动以来的单调时间戳（毫秒）
- * @retval 当前毫秒时间戳（约 49.7 天溢出，嵌入式场景可接受）
+ * @retval 当前毫秒时间戳（uint64_t，无实际溢出风险）
  */
-uint32_t time_util_get_ms(void);
+uint64_t time_util_get_ms(void);
 
 /**
  * @brief  计算两个时间戳之间的经过毫秒数
  * @param  start_ms  起始时间戳（由 time_util_get_ms() 获取）
  * @param  now_ms    当前时间戳（由 time_util_get_ms() 获取）
- * @retval 经过的毫秒数
- * @note   必须用此函数计算时间差，禁止用 >= 直接比较两个时间戳——
- *         uint32_t 无符号减法对回绕（约 49.7 天溢出）天然正确，
- *         而直接比较在溢出后会得到错误结果。
+ * @retval 经过的毫秒数（业务间隔远小于 UINT32_MAX，返回值保持 uint32_t）
+ * @note   必须用此函数计算时间差，禁止用 >= 直接比较两个时间戳。
  */
-static inline uint32_t time_elapsed_ms(uint32_t start_ms, uint32_t now_ms)
+static inline uint32_t time_elapsed_ms(uint64_t start_ms, uint64_t now_ms)
 {
-    return now_ms - start_ms;
+    return (uint32_t)(now_ms - start_ms);
 }
 
 #ifdef __cplusplus
